Clarify widget settings being available only when using proxy.
The Connect Widget's settings are documented in the README, but it was not
clear that they were only applied when using the proxy. This changes that by
moving the documentation for those settings into the proxy section and being
explicitly about them only working when the proxy is in use.